Store 사용 후 Status와 Furniture 업데이트 되게
