+++ Music Is My Life/Assets/Scripts/Main-Scripts/MainUpdateController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f4fe81c5948e6c4caaf0cb99a19f7cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
