+++ Music Is My Life/Assets/Scripts/Main-Scripts/StatusController.cs
@@ -11,58 +11,14 @@ public class StatusController : MonoBehaviour
     public GameObject[] allAppsObj; // 알바만 넣었음
     public GameObject floorThing;
 
-    void Start()
+    public void UpdateStatus()
     {
-        StatusChanger.UpdateDay(); // 날짜 업데이트
-        // 14일이 지나면 엔딩으로 이동
-        if (PlayerPrefs.GetInt("Dday") > 14)
-        {
-            GoToEnding();
-        }
-        // 스트레스가 100 이상일 경우 게임오버
-        if (PlayerPrefs.GetInt("Stress") >= 100)
-        {
-            GoToGameOver();
-        }
         NameText(); // 플레이어 이름
         DdayText();
         DateText();
         StatusText();
         SetAppsByYesterdayBehavior();
         SetFloorThing(floorThing);
-    }
-
-    public void GoToEnding() // 엔딩으로
-    {
-        int money = PlayerPrefs.GetInt("Money");
-        int myFame = PlayerPrefs.GetInt("MyFame");
-        int bandFame = PlayerPrefs.GetInt("Fame");
-
-        SceneMove sceneMove = gameObject.AddComponent<SceneMove>();
-        if (money > 2500000)
-        {
-            sceneMove.targetScene = "Ending-Expedition";
-        }
-        else if (myFame > 100)
-        {
-            sceneMove.targetScene = "Ending-OpeningBand";
-        }
-        else if (bandFame > 15)
-        {
-            sceneMove.targetScene = "Ending-ConcertInKorea";
-        }
-        else
-        {
-            sceneMove.targetScene = "Ending-Normal";
-        }
-        sceneMove.ChangeScene();
-    }
-
-    public void GoToGameOver()
-    {
-        SceneMove sceneMove = gameObject.AddComponent<SceneMove>();
-        sceneMove.targetScene = "Ending-GameOver";
-        sceneMove.ChangeScene();
     }
 
     public void NameText()
